Allow expenses to be not counted against budget
They will still be recorded as expenses (and the total expenses report
will count them), but the "daily spending cap" won't be affected by
them. This is useful for the "current month", but poses problems later
when the refund comes in.

The refund should also be *not* counted against budget and this quicly
leads to a messy situation when the revenue/expense calculations are not
exact and can be wildly off if a "not in budget" tag is forgotten.
Imagine that one month you don't count a 1000 PLN expense against
budget, and the next month when the refund comes you forget to mark it
correctly. Now the reports says that you have an extra 1000 PLN to spend
even if the money was *already spent* last month.

I guess this is a bad feature and should not be added.

Maybe a "count against budget of YYYY-MM" would be better so the expense
is recorded in the month when it occured, but would count against budget
of the month when the money for it would come as revenue.