Allow overriding values used in calculations of expenses
An example:

    tx 1971-01-01T00:01 Test exchange for "Text expense"
        asset/bank.pln  -180.00 PLN
        asset/bank.main   40.00 EUR
    with
        rate: EUR/PLN 4.5
    end
    ex 1971-01-01T00:02 Test expense
        asset/bank.main -40.00 EUR
        Some Company, Inc.
    with
        calculate_as: -180.00 PLN
    end

This serves as a tool to get more accurate estimates of spent money when
the expense is in a different currency than the default.

Let's say that you were buying the foreign currency for quite some time
and the ledger knows the weighted average price you paid. It would use
this value to calculate your expense. However, if you just bought the
foreign currency and immediately spent it to pay the bill then the
calculation would be wrong - you know that the transfer you entered into
the ledger is the real price, but the software does not. So while it
gives the correct result (in theory), it is kind of wrong (in practice).
The "calculate_as" attribute is the way to alleviate that.