Support uniform currency rate specification
The default currency no longer needs to be specified first. The software
will recalculate the rate internally to the "default-currency-first"
format, but the user may enter it however they wish.

The old format was:

    rate: DEFAULT/SOME 2.5

which meant that "I have to pay 2.5 DEFAULT to get 1 SOME". However, if
the rate on the receipt was specified "SOME/DEFAULT" you would have to
manually change the format so that rates were always specified as
"DEFAULT/SOME".

Now, the software understands both ways and treats "FOO/BAR 2.5" as "I
have to pay 2.5 BAR to get 1 FOO". In practice these two rates are the
same:

    rate: PLN/EUR 0.2166
    rate: EUR/PLN 4.6168