Make the reports multi-column to better utilise screen real estate
Just like in the old ledger, the report is now spread into several
columns. The first (leftmost) is reserved for "This period" (eg, this
day, week, month, year); the second (middle) is reserved for the "Last
period" (eg, last day, week, month, year); the third is occupied by some
other period - eg, "all time" or previous 14 days.

This makes the report more compact and increases information density as
it is packing the same amount of info using less screen space.