Fix receiving dividends in currency different from share price
This case came up when I received a dividend paid in NOK for shares
priced in EUR. At the current exchange rate of about 11 NOK for 1 EUR
the total return from the shares shot up into hundres of percent, which
was ridiculous.

The ledger must detect when the dividend is paid in a currency that does
not match the share price and adjust accordingly. Of course, precise
calculations rely on the exchange rates being at least somewhat recent.