Fix loading currency exchange rates
Assuming a pair of currencies FOO and BAR, with 1 FOO being worth 2 BAR,
the exchange rate could be set like this:

    FOO/BAR 2.0000

or like this:

    BAR/FOO 0.5000

However, before this patch this would not work. Setting FOO/BAR would
only overwrite the rate set by a previous FOO/BAR and not any BAR/FOO.
This also meant that there could be two exchange rates active at the
same time for a single pair.