Release 0.5.0
TOTAL RETURN is not displayed in its own column anymore. It is now the
default value displayed for gain/loss from a company in the equity
report.

The default period report includes information on EXPENSE SINKS and
REVENUE FAUCETS. This lets you more easily monitor where does your money
flows in from, and where does it flow out to.