Add backgrounds to sinks and expenses
This visualises their relative weights in a more graphic way. The raw
percentage is still visible, but the bars are a nice supporting effect.

The bars for revenue faucets are scaled using VISIBILITY_MULTIPLIER to
make the second and following ones wider. Why? Because for people on
salary who are employed by a single company (including yours truly) the
main faucet usually completely OVERWHELMS the other sources.

This resulted in the "Faucet" section only having one big bar, and maybe
a tiny one on the second faucet.

The bars serve a secondary function: if a few of the top faucet bars are
rougly equal you can quickly judge your situation as pretty good. It
means that you do not depend on any single source on income... that
much.