Base cash flow report on net values
This is explained by the comment I put over the sinks, so I will
reproduce it here:

  Base values for expenses and revenues can be either the total (ie, every
  expense incurred and revenue earned during the analysed period) or the net
  value. The net value is calculated by summing expenses and revenues and
  only considering the final outcome.

  Sinks are ALREADY net so reporting precentages makes more sense based on
  the net value.

  Keep in mind that this distinction (total vs net) is only meaningful and
  affects calculations for entities that are both expense sinks and revenue
  faucets. To give a couple of more concrete examples:

   1/ Assume that you are renting a flat. Each month you must pay the
      landlord a certain amount of cash. The landlord is a pure expense sink
      and WILL NOT be affected by the net/total distintion.
   2/ You go to work and receive a paycheck every month. The employer is a
      pure revenue faucet.
   3/ You and your mother have a birthday the same month, and you both gift
      each other $100. In this case the mother is BOTH an expense sink (as
      you have given her $100) and a revenue faucet (she has given you $100).
      The total cash flow is +$100 and -$100, but the net flow is $0.
      Because there is a sink and a faucet for your mother, she WILL be
      affected by the net/total distintion.

This notion was already present in the code (ie, see the code that
"nets" the expense sinks and revenue faucets) but was not brought to its
conclusion.

The effect was that sink and faucet reports were misleading: the nominal
values were NET, but the percentages were based on TOTAL flows. If you
tried adding up either nominals or percentages you would not get a
satisfactory result matching with reported Revenues, Expenses, and Net.

This is now all fixed.

The result of the patch is a report that more accurately describes
reality. If you had mixed sinks/faucets in your cash flow you will see a
difference (although it should be relatively minor on the yearly and all
time reports).