Improve equity accounts report
Add a column which states the weight of a position ie, the percentage of
total portfolio value that the position is worth. This is based on
current market value of the position so only share market price is
considered.

Also, display legends above and below the report to make it immediately
obvious what the different columns represent:

 - Market Price: last price for which the share was bought/sold on the
   market
 - Average Price: average price of a share in portfolio
 - Share Count: number of shares held
 - Market Value: value of shares on the market ie, number of shares held
   times the market price
 - Port%: value of the position as percentage of the total portfolio
   value
 - Total Return (TR$): total return as a value in share's currency
 - Total Return (TR%): total return as a percentage

Total return is calculated as share price appreciation plus any
dividends received versus cost basis of the position.

The report is most useful for people with a strategy similar to mine,
which is pretty simple: buy and hold.