Provide a way to aggregate sinks and faucets
Tracking spending by sink is the reasonable way to do it, I think. For
example, if you refuel your car you sometimes do it on OMV, sometimes on
Eni, and sometimes on Shell. However, it would be more useful to just
see a "Fuel" position on the report.

To continue the above example:

    group Fuel
        OMV
        Eni
        Shell
    end

With the above text present in your ledger, the expenses (and revenues)
from OMV, Eni, and Shell will be reported as [[Fuel]] position.